+++ src/ProcedureOperation.cpp
@@ -8,7 +8,7 @@ namespace mssql
 {
 	ProcedureOperation::ProcedureOperation(shared_ptr<OdbcConnection> connection, 
 		const wstring& query, 
-		u_int id, 
+		size_t id, 
 		u_int timeout, 
 		Handle<Object> callback) :
 		QueryOperation(connection, query, id, timeout, callback)
