+++ src/QueryOperation.h
@@ -33,7 +33,7 @@ namespace mssql
 	class QueryOperation : public OdbcOperation
 	{
 	public:
-		QueryOperation(shared_ptr<OdbcConnection> connection, const wstring& query, u_int queryId, u_int timeout, Handle<Object> callback);
+		QueryOperation(shared_ptr<OdbcConnection> connection, const wstring& query, size_t queryId, u_int timeout, Handle<Object> callback);
 		bool BindParameters(Handle<Array> & node_params) const;
 		Local<Array> UnbindParameters() const;
 		// called by BindParameters when an error occurs.  It passes a node.js error to the user's callback.
